adds git plugin, removes vscode config, better kitty
the git messenger plugin is very cool. It adds a floating window with
info around the commit that introduced a line. Might have to add it to
the leader mapper menus in case I forget about it.

I haven't used VSCode in a really long time so I thought I might as well
get rid of the config. The plugin ecosystem has probably changed anyway.

I also have started to use kitty more. Alacritty was really slow with
tmux and nvim running, and kitty has been prettier and faster, which I
like so far. I had to add a few small things to the config but nothing
major.