moves to new diagnostic mappings and adds in wiki vim
The Wiki is my white whale if ever there was one. I just can't get
settled into one and I don't know why. I will eventually be forced to
write one from scratch, and take up all of my time. This is the way.

The tricky thing is that I fundamentally don't yet have all my
requirements. I must remember that that is ok, and good, and the way of
things. But still...

I simply wish that I knew if I needed the wiki to end up online, or if
I'll need to be able to edit on my phone. Will I write a blog from my
wiki? Surely not, but still I can dream.