setup luasnip with some pretty cool snips
luasnip lets you do some pretty cool stuff, and the api is actually
pretty great once you get used to it. The only thing that's tough isn't
because of luasnip, it's because of the cmp integration. Postfix
snippets aren't selectable easily from the cmp menu, and they come out
weird when they are.